feat(content): update academic information in index.md
This commit adds the link to the C.V. and creates a new section 'Academic things' in the index.md file. It further rephrases the existing line about research project summaries, providing it under the newly created section. Now users can directly access the C.V., improving the accessibility of academic credentials on the page.